+++ src/main/java/bugztracker/repository/IProjectRepository.java
@@ -13,7 +13,7 @@ public interface IProjectRepository extends IRepository<Project> {
     List<Project> search(String text);
     List<Project> getSortedList(String nameField, String option);
 
-    Project getProjectWithUsers(long id);
+    Project getProjectWithUsers(int id);
 
     List<Project> getProjectsOfUser(User user);
 }
